+++ package.json
@@ -10,10 +10,12 @@
         "transpiler"
     ],
     "scripts": {
-        "build": "tsc -p tsconfig.json",
-        "test": "gulp test",
-        "test-threaded": "tslint -p . && ts-node ./test/threaded_runner.ts",
-        "coverage-threaded": "nyc npm run test-threaded && nyc report --reporter=text-lcov > coverage.lcov",
+        "build": "tsc -p tsconfig.json && npm run build-lualib",
+        "build-lualib": "ts-node ./build_lualib.ts",
+        "test": "tslint -p . && npm run build && ts-node ./test/runner.ts",
+        "coverage": "nyc npm test && nyc report --reporter=text-lcov > coverage.lcov",
+        "coverage-html": "nyc npm test && nyc report --reporter=html",
+        "test-threaded": "tslint -p . && npm run build && ts-node ./test/threaded_runner.ts",
         "release-patch": "npm version patch",
         "release-minor": "npm version minor",
         "release-major": "npm version major",
@@ -30,6 +32,9 @@
         ],
         "include": [
             "src/**/*"
+        ],
+        "exclude": [
+            "src/lualib/*"
         ]
     },
     "engines": {
@@ -40,29 +45,17 @@
         "yargs": "^11.1.0"
     },
     "devDependencies": {
-        "@types/del": "^3.0.1",
         "@types/glob": "^5.0.35",
-        "@types/gulp": "^4.0.5",
-        "@types/gulp-concat": "0.0.32",
-        "@types/gulp-istanbul": "^0.9.32",
-        "@types/gulp-shell": "0.0.31",
         "@types/node": "^9.6.23",
         "@types/yargs": "^11.0.0",
         "alsatian": "^2.2.1",
         "circular-json": "^0.5.5",
         "codecov": "3.0.2",
+        "concat": "^1.0.3",
         "deep-equal": "^1.0.1",
-        "del": "^3.0.0",
         "fengari": "^0.1.2",
         "glob": "^7.1.2",
-        "gulp": "^4.0.0",
-        "gulp-concat": "^2.6.1",
-        "gulp-istanbul": "^1.1.3",
-        "gulp-shell": "^0.6.5",
-        "gulp-tslint": "^8.1.3",
-        "gulp-typescript": "^5.0.0-alpha.3",
         "nyc": "^11.9.0",
-        "remap-istanbul": "^0.11.1",
         "threads": "^0.12.0",
         "ts-node": "^7.0.0",
         "tslint": "^5.10.0",
