+++ src/Compiler.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 import * as fs from "fs";
 import * as path from "path";
 import * as ts from "typescript";
