+++ src/Index.ts
@@ -1,3 +1,5 @@
+#!/usr/bin/env node
+
 import {compile} from "./Compiler";
 
 compile(process.argv.slice(2));
